backend/requirements.txt
beautifulsoup4==4.12.3
websearch-python==1.2.2
uvicorn==0.27.1
fastapi==0.110.0
accelerate==0.27.2
tokenizers==0.15.2
bitsandbytes==0.43.0
einops==0.7.0
faiss-cpu==1.8.0
transformers==4.37.2
sentence_transformers==2.2.2
langchain==0.1.11
transformers_stream_generator==0.0.5
protobuf==3.20.3